added memory available and used graphs

+++ logging_agent/src/commonMain/kotlin/com/zoho/vtouch/logging_agent/model/GraphData.kt
@@ -3,5 +3,5 @@ package com.zoho.vtouch.logging_agent.model
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class GraphData(val id:String, val value: Int, val timestamp: Long)
+data class GraphData(val id:String, val value: Float, val timestamp: Long)
